% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgBoxWidget.R
\name{msgBoxWidget}
\alias{msgBoxWidget}
\title{msgBoxWidget}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing
}
\description{
an R6 class providing clean access to a simple html message box
}
\details{
R6 Class representing a simple text display widget
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{msgBoxWidget$new()}}
\item \href{#method-ui}{\code{msgBoxWidget$ui()}}
\item \href{#method-server}{\code{msgBoxWidget$server()}}
\item \href{#method-clone}{\code{msgBoxWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new msgBoxWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msgBoxWidget$new(
  id,
  title,
  boxWidth = 200,
  boxHeight = 30,
  fontSize = 20,
  fontColor = "black",
  backgroundColor = "beige"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{title}}{character}

\item{\code{boxWidth}}{integer, 200 by default}

\item{\code{boxHeight}}{integer, 30 by default}

\item{\code{fontSize}}{integer, 20 by default}

\item{\code{fontColor}}{character, standard CSS rgb or name}

\item{\code{backgroundColor}}{character, standard CSS rgb or name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `msgBoxWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msgBoxWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msgBoxWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msgBoxWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
